requirements.txt
fastapi
uvicorn
pandas==3.0.0
numpy==2.4.2
scipy==1.17.0
pymongo==4.16.0
dnspython==2.8.0
torch
faiss-cpu
tqdm
python-dotenv

